# internal/infra/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package db

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Brand struct {
	ID          uuid.UUID
	Name        string
	Description sql.NullString
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type Category struct {
	ID          uuid.UUID
	Name        string
	Description sql.NullString
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type Coupon struct {
	ID             uuid.UUID
	Code           string
	Discount       string
	ExpirationDate time.Time
	Status         string
	DeletedAt      sql.NullTime
	CreatedDate    sql.NullTime
	UpdatedDate    sql.NullTime
}

type Customer struct {
	ID          uuid.UUID
	Name        string
	Email       string
	Password    string
	Address     sql.NullString
	Phone       sql.NullString
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type Delivery struct {
	ID             uuid.UUID
	OrderID        uuid.NullUUID
	CustomerID     uuid.NullUUID
	DeliveryType   string
	DeliveryDate   sql.NullTime
	DeliveryStatus string
	DeletedAt      sql.NullTime
	CreatedDate    sql.NullTime
	UpdatedDate    sql.NullTime
}

type Feedback struct {
	ID          uuid.UUID
	CustomerID  uuid.NullUUID
	OrderID     uuid.NullUUID
	Rating      sql.NullInt32
	Comment     sql.NullString
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type Order struct {
	ID          uuid.UUID
	CustomerID  uuid.NullUUID
	OrderDate   sql.NullTime
	OrderStatus string
	TotalValue  string
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type OrderItem struct {
	ID        uuid.UUID
	OrderID   uuid.NullUUID
	ProductID uuid.NullUUID
	Quantity  int32
	UnitPrice string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type OrderProduct struct {
	ID          uuid.UUID
	OrderID     uuid.NullUUID
	ProductID   uuid.NullUUID
	Quantity    int32
	UnitPrice   string
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}

type Payment struct {
	ID            uuid.UUID
	OrderID       uuid.NullUUID
	CustomerID    uuid.NullUUID
	PaymentType   string
	PaymentDate   sql.NullTime
	PaymentValue  string
	PaymentStatus string
	DeletedAt     sql.NullTime
	CreatedDate   sql.NullTime
	UpdatedDate   sql.NullTime
}

type Product struct {
	ID              uuid.UUID
	Name            string
	Description     sql.NullString
	Price           string
	QuantityInStock int32
	ImageUrl        sql.NullString
	CategoryID      uuid.NullUUID
	BrandID         uuid.NullUUID
	DeletedAt       sql.NullTime
	CreatedDate     sql.NullTime
	UpdatedDate     sql.NullTime
}

type ProductStock struct {
	ID        uuid.UUID
	ProductID uuid.NullUUID
	StockID   uuid.NullUUID
	Quantity  int32
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type Stock struct {
	ID          uuid.UUID
	Name        string
	Location    sql.NullString
	Capacity    int32
	DeletedAt   sql.NullTime
	CreatedDate sql.NullTime
	UpdatedDate sql.NullTime
}